Moved transformTsqResultsForVisualization to direct import transformers
